Moved util to root package.
