+++ govcd/catalog.go
@@ -9,8 +9,8 @@ import (
 	"encoding/xml"
 	"errors"
 	"fmt"
-	"github.com/vmware/go-vcloud-director/govcd/util"
 	"github.com/vmware/go-vcloud-director/types/v56"
+	"github.com/vmware/go-vcloud-director/util"
 	"io"
 	"io/ioutil"
 	"log"
